+++ crates/scheduler/src/call/crypto.rs
@@ -14,6 +14,7 @@
 
 mod ccm;
 mod gcm;
+mod hash;
 
 use wasefire_applet_api::crypto::Api;
 use wasefire_board_api::Api as Board;
@@ -24,5 +25,6 @@ pub fn process<B: Board>(call: Api<DispatchSchedulerCall<B>>) {
     match call {
         Api::Ccm(call) => ccm::process(call),
         Api::Gcm(call) => gcm::process(call),
+        Api::Hash(call) => hash::process(call),
     }
 }